Add files via upload
A perfect number is a positive integer that is equal to the sum of its proper divisors. For example, 6, which is the sum of 1, 2, and 3.

Logic :
To check if the number is perfect or not we have to run one loop from 1 to N and sum all the numbers between 1 to N, if the sum is equal to N then it is a perfect number.